re-configured login/register and added some css for cart

+++ src/components/Authenitcation/Login.js
@@ -3,17 +3,19 @@
 //~~~~~~~~~~~~~~~~~~~
 
 import React, { useState } from 'react';
-import { loginUser } from '../../api/users';
-import { NavLink, Redirect } from 'react-router-dom';
+import { checkUserByUsername, loginUser } from '../../api/users';
+import { NavLink, useHistory } from 'react-router-dom';
 import './Auth.css';
 
-const Login = ({ currentUser, setCurrentUser }) => {
+const Login = ({ setCurrentUser }) => {
   //~~~~~~~~~~~~~~~~~~~
   //~~~~~~ STATE ~~~~~~
   //~~~~~~~~~~~~~~~~~~~
 
   const [username, setUsername] = useState('');
   const [password, setPassword] = useState('');
+
+  let history = useHistory();
 
   async function handleSubmit(event) {
     event.preventDefault();
@@ -22,25 +24,38 @@ const Login = ({ currentUser, setCurrentUser }) => {
       alert('Please enter a valid username or password');
       return;
     }
+    try {
+      const userUsername = await checkUserByUsername({ username });
+      console.log(userUsername);
 
-    const user = await loginUser({ username, password });
-    console.log(user);
+      // whis is this if statement not working
+      if (userUsername === undefined) {
+        alert('Username does not exist');
+        return;
+      }
+      const user = await loginUser({ username, password });
 
-    localStorage.clear();
-    //When logging in after changing username and password user comes back as undefined.
-    //I suspect a bcrypt problem.
-    setCurrentUser({
-      id: user.user.id,
-      username: user.user.username,
-      admin: user.user.admin,
-    });
+      console.log(user);
 
-    localStorage.setItem('id', user.user.id);
-    localStorage.setItem('username', user.user.username);
-    localStorage.setItem('token', user.token);
+      localStorage.clear();
+      //When logging in after changing username and password user comes back as undefined.
+      //I suspect a bcrypt problem.
+      setCurrentUser({
+        id: user.user.id,
+        username: user.user.username,
+        admin: user.user.admin,
+      });
 
-    setUsername('');
-    setPassword('');
+      localStorage.setItem('id', user.user.id);
+      localStorage.setItem('username', user.user.username);
+      localStorage.setItem('token', user.token);
+
+      setUsername('');
+      setPassword('');
+      history.push('/store');
+    } catch (error) {
+      throw error;
+    }
   }
 
   //~~~~~~~~~~~~~~~~~~~
@@ -61,48 +76,40 @@ const Login = ({ currentUser, setCurrentUser }) => {
 
   return (
     <div id='login'>
-      {currentUser.id === 1 ? (
-        <>
-          <form onSubmit={handleSubmit} className='auth-form'>
-            <h1 className='auth-header'>Welcome</h1>
-            <div className='auth-box'>
-              <label htmlFor='username' className='auth-label'>
-                Username:
-              </label>
-              <input
-                className='auth-input'
-                type='text'
-                name='username'
-                value={username}
-                onChange={handleUsernameChange}
-              />
-            </div>
-            <div className='auth-box'>
-              <label htmlFor='password' className='auth-label'>
-                Password:
-              </label>
-              <input
-                className='auth-input'
-                type='password'
-                name='password'
-                value={password}
-                onChange={handlePasswordChange}
-              />
-            </div>
-            <button className='auth-button' type='submit'>
-              Login
-            </button>
-            <div className='accountCheck'>
-              <h2 className='memberCheck'>Not A Mac Member Yet?</h2>
-              <NavLink to='/register'>REGISTER</NavLink>
-            </div>
-          </form>
-        </>
-      ) : (
-        <>
-          <Redirect to='/home' />
-        </>
-      )}
+      <form onSubmit={handleSubmit} className='auth-form'>
+        <h1 className='auth-header'>Welcome</h1>
+        <div className='auth-box'>
+          <label htmlFor='username' className='auth-label'>
+            Username:
+          </label>
+          <input
+            className='auth-input'
+            type='text'
+            name='username'
+            value={username}
+            onChange={handleUsernameChange}
+          />
+        </div>
+        <div className='auth-box'>
+          <label htmlFor='password' className='auth-label'>
+            Password:
+          </label>
+          <input
+            className='auth-input'
+            type='password'
+            name='password'
+            value={password}
+            onChange={handlePasswordChange}
+          />
+        </div>
+        <button className='auth-button' type='submit'>
+          Login
+        </button>
+        <div className='accountCheck'>
+          <h2 className='memberCheck'>Not A Mac Member Yet?</h2>
+          <NavLink to='/register'>REGISTER</NavLink>
+        </div>
+      </form>
     </div>
   );
 };
